// src/app/services/swagger/model/confirmedMove.ts
/**
 * BoF REST API
 * Battle for honor interface
 *
 * OpenAPI spec version: 0.1.4
 * Contact: webmaster@batleforhonor.de
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { FleetMarker } from './fleetMarker';
import { Move } from './move';

/**
 * .
 */
export interface ConfirmedMove { 
    /**
     * The fleets which are confirmed for the move.
     */
    attendants: Array<FleetMarker>;
    move: Move;
    /**
     * The move identifier.
     */
    moveHash: number;
}